requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
mysql-connector-python==8.2.0
itsdangerous==2.1.2
werkzeug==3.0.1
sqlalchemy==2.0.23
pytesseract==0.3.10
pdf2image==1.16.3
cohere==4.37
transformers==4.35.2
requests==2.31.0
PyPDF2==3.0.1
pdfminer.six==20221105
pydub==0.25.1
Pillow==10.1.0
gtts==2.4.0
pymysql==1.1.0
gunicorn==21.2.0
